convert input files in 40bit long int values
this c code first converts the input values to 40bit long values. we are using 40 bits here because we know that all the computations will result in 40 bit values, so we are converting the inputs to 40 bits using trailing zeroes. 
Also note that the code/computation is done in hexadecimal format (eg: 0xFF00000000) for sancity of coding.